Rename plugins to inputs
Since Telegraf works with the concept of plugins for both inputs and
outputs, making it explicit which kind of plugin are we handling in each
section is important to avoid mistakes and confusions.

After this commit I intend to improve this ansible role in order to
allow multiple output plugins also, which will also make 'influxdb'
optional as output.